userspace: split address representability checks out of policies.go
Move the address representability concern verbatim from policies.go into a
new policies_representable.go: allAddressTokensRepresentable,
nameRepresentable, nameRepresentability, addressValueRepresentable. These
decide whether the userspace matcher can represent a policy address token /
address-book name (#3261/#3294) — the structural parity contract with the
strict commit validator.

The low-level CIDR predicates isV4CIDR/isV6CIDR stay in the residual
policies.go: they are shared between this concern (addressValueRepresentable)
and the address-book concern (expandBookNameToCIDRs), so they belong to the
shared core rather than either sibling.

Pure code-motion, no logic change: byte-identical slices of the
origin/master policies.go. Refs #4421 (codex-173 #4).
